Add files via upload
In this end-to-end data analytics project, we will take a REAL dataset for employee presence from a company called AtliQ and perform data analysis in Power BI. We invited an HR from AtliQ Technologies who conveyed her requirements and we built a Power BI dashboard for her to help her with employee insights. This project is perfect for your data analyst resume as it is built on a real dataset (and not the toy dataset).